Move FieldSeparator to ColumnSeparator, fix JavaDoc
